Moving http request to a dedicated file
