+++ WikiCat/AppIntent/RandomCatBreedIntent.swift
@@ -19,20 +19,20 @@ struct RandomCatBreedIntent: AppIntent {
     
     @MainActor
     func perform() async throws -> some IntentResult & ShowsSnippetView {
-        let modelData = ModelData()
+        let requestManager = HttpRequestManager.shared
         var referenceImage: UIImage? = nil
         
         // Get a random CatBreed
-        await modelData.fetchBreeds()
+        await requestManager.fetchBreeds()
         // Wait a second to be sure the breed is retrieved
         try await Task.sleep(nanoseconds: UInt64(1.2 * Double(NSEC_PER_SEC)))
-        guard let catBreed = modelData.breedsList.randomElement() else {
+        guard let catBreed = ModelData.shared.breedsList.randomElement() else {
             throw MyIntentError.message("We couldn't retrieve a random breed 😿")
         }
         
         // Get a random CatImage and download it
         if catBreed.image != nil, let url = URL(string: catBreed.image!.url) {
-            modelData.downloadImage(from: url) { data, response, error in
+            requestManager.downloadImage(from: url) { data, response, error in
                 guard let data = data, error == nil else { return }
                 referenceImage = UIImage(data: data)
             }
@@ -42,7 +42,7 @@ struct RandomCatBreedIntent: AppIntent {
             // If CatImage is nil in CatBreed check if there is a reference_id
             let url = URL(string: "https://cdn2.thecatapi.com/images/\(catBreed.reference_image_id!).jpg")
             if let url = url {
-                modelData.downloadImage(from: url) { data, response, error in
+                requestManager.downloadImage(from: url) { data, response, error in
                     guard let data = data, error == nil else { return }
                     referenceImage = UIImage(data: data)
                 }
